requirements.txt
python-telegram-bot==21.5
anthropic==0.34.2
apscheduler==3.10.4
httpx==0.27.2
pytz==2024.1
